+++ weasyl/controllers/api.py
@@ -4,7 +4,7 @@ import web
 from libweasyl.text import markdown, slug_for
 from libweasyl import ratings
 
-from weasyl.controllers.base import controller_base
+from weasyl.controllers.decorators import controller_base
 from weasyl.error import WeasylError
 from weasyl import define as d, macro as m
 from weasyl import (
